Struct ncollide3df64​::geom​::Cone​

pub struct Cone {
    // some fields omitted
}

Implicit description of a cylinder geometry with its principal axis aligned with the y axis.

Methods

impl Cone

fn new(half_height: Scalar, radius: Scalar) -> Cone

Creates a new cone.

Arguments:

half_height - the half length of the cone along the y axis.

radius - the length of the cone along all other axis.

fn new_with_margin(half_height: Scalar, radius: Scalar, margin: Scalar) -> Cone

Creates a new cone with a custom margin.

Arguments:

half_height - the half length of the cone along the y axis.

radius - the length of the cone along all other axis.

margin - the cone margin.

impl Cone

fn half_height(&self) -> Scalar

The cone half length along the y axis.

fn radius(&self) -> Scalar

The radius of the cone along all but the y axis.

fn margin(&self) -> Scalar

The margin around the cone.