Struct ncollide3df64​::geom​::GeomWithMargin​

pub struct GeomWithMargin<'a, G> {
    // some fields omitted
}

Extends the wrapped geometry with its margin.

The purpose of this structure as to make the support_point_without_margin method return result of the wrapped geometry support_point method instead.

Methods

impl<'a, G> GeomWithMargin<'a, G>

fn new(geom: &'a G) -> GeomWithMargin<'a, G>

Creates a new GeomWithMargin.

fn geom<'b>(&'b self) -> &'b G

The wrapped geometry.