Trait ncollide3df64​::geom​::MeshElement​

pub trait MeshElement {
    fn nvertices(unused: Option<Self>) -> uint;
    fn new_with_vertices_and_indices(&[Vect], &[uint], Scalar) -> Self;
}

Trait implemented by elements usable on the Mesh.

Note that this trait is not useful per se since the Mesh is not parameterized by the element type. However, types implementing this trait is valid as a type alias for MeshPrimitive.

Required Methods

fn nvertices(unused: Option<Self>) -> uint

The number of vertices of this mesh element.

fn new_with_vertices_and_indices(&[Vect], &[uint], Scalar) -> Self

Creates a new mesh element from a set of vertices and indices and the margin.