Trait ncollide3df64​::implicit​::PreferedSamplingDirections​

pub trait PreferedSamplingDirections<Vect, Matrix> {
    fn sample(&self, &Matrix, |Vect|);
}

Trait of geometries having prefered sampling directions for the Minkowski sampling algorithm.

Those directions are usually the geometry faces normals.

Required Methods

fn sample(&self, &Matrix, |Vect|)

Applies a function to this geometry with a given transform.