Trait ncollide3df64​::parametric​::ParametricSurface​

pub trait ParametricSurface {
    fn at(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_u(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_v(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_uu(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_vv(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_uv(&self, u: Scalar, v: Scalar) -> Vect;
    fn at_uv_nk(&self, u: Scalar, v: Scalar, n: uint, k: uint) -> Vect;

    fn at_u_v(&self, u: Scalar, v: Scalar) -> (Vect, Vect, Vect) { ... }
    fn at_u_v_uu_vv_uv(&self, u: Scalar, v: Scalar) -> (Vect, Vect, Vect, Vect, Vect, Vect) { ... }
    fn at_uv_nk_all(&self, u: Scalar, v: Scalar, n: uint, k: uint, out: &mut DMat<Vect>) { ... }
}

Trait implemented by differentiable parametric surfaces.

The parametrization space is assumed to be [0.0, 1.0] x [0.0, 1.0].

Required Methods

fn at(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the parametric surface.

fn at_u(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the surface derivative wrt. u.

fn at_v(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the surface derivative wrt. v.

fn at_uu(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the surface second derivative wrt. u.

fn at_vv(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the surface second derivative wrt. v.

fn at_uv(&self, u: Scalar, v: Scalar) -> Vect

Evaluates the surface second derivative wrt. u and v.

fn at_uv_nk(&self, u: Scalar, v: Scalar, n: uint, k: uint) -> Vect

Evaluates the parametric surface and its derivative wrt. u n times and wrt. v k times.

Returns dn + k S(u, v) / (dun dkn).

Provided Methods

fn at_u_v(&self, u: Scalar, v: Scalar) -> (Vect, Vect, Vect)

Evaluates the parametric surface and its first derivatives.

Returns (S(u, v), dS / du, dS / dv)

fn at_u_v_uu_vv_uv(&self, u: Scalar, v: Scalar) -> (Vect, Vect, Vect, Vect, Vect, Vect)

Evaluates the parametric surface and its first and second derivatides.

Returns (S(u, v), dS / du, dS / dv, d²S / du², d²S / dv², d²S / dudv)

fn at_uv_nk_all(&self, u: Scalar, v: Scalar, n: uint, k: uint, out: &mut DMat<Vect>)

Evaluates all the partial derivatives of the surface, up to the n-th derivative.

The output matrix will contain the derivatives dn S(u, v) / (dui dvj) at the cells indexed by (i, j).