Module ncollide3df64​::procedural​

Procedural mesh generation.

Reexports

pub use procedural::trimesh::{UnifiedIndexBuffer, SplitIndexBuffer};

Modules

path
Path generation.

utils
Utilities useful for various generations tasks.

Structs

Polyline
Geometric description of a polyline.

TriMesh
Geometric description of a mesh.

Enums

IndexBuffer
Different representations of the index buffer.

Traits

ToPolyline
Trait implemented by geometries that can be approximated by a triangle mesh.

ToTriMesh
Trait implemented by geometries that can be approximated by a triangle mesh.

Functions

bezier_curve
Given a set of control points, generates a (non-rational) Bezier curve.

bezier_surface
Given a set of control points, generates a (non-rational) Bezier surface.

capsule
Generates a capsule.

circle
Creates a circle lying on the (x,y) plane.

cone
Generates a cone with a given height and diameter.

convex_hull2d
Computes the convex hull of a set of 2d points.

convex_hull2d_idx
Computes the convex hull of a set of 2d points and returns only the indices of the hull vertices.

convex_hull3d
Computes the convex hull of a set of 3d points.

cuboid
Generates a cuboid geometry with a split index buffer.

cylinder
Generates a cylinder with a given height and diameter.

hacd
Approximate convex decomposition of a triangle mesh.

parametric_surface_uniform
Meshing algorithm that uniformly triangulates the parametric space.

parametric_surface_uniform_with_distance_error
Uniformly triangulates the parametric space.

quad
Adds a double-sided quad to the scene.

quad_with_vertices
Adds a double-sided quad with the specified grid of vertices.

rational_bezier_curve
Given a set of control points, generates a rational Bezier curve.

rational_bezier_surface
Given a set of weighted control points, generates a rational Bezier surface.

rectangle
The contour of a cuboid lying on the x-y plane.

sphere
Generates a UV sphere.

unit_circle
Creates a circle lying on the (x,y) plane.

unit_cone
Generates a cone with unit height and diameter.

unit_cuboid
Generates a cuboid geometry with a split index buffer.

unit_cylinder
Generates a cylinder with unit height and diameter.

unit_quad
Adds a double-sided quad with unit size to the scene.

unit_rectangle
The contour of a unit cuboid lying on the x-y plane.

unit_sphere
Generates a UV sphere centered at the origin and with a unit diameter.