Trait ncollide3df64​::procedural​::path​::StrokePattern​

pub trait StrokePattern<N, V> {
    fn stroke<C: CurveSampler<N, V>>(&mut self, sampler: &mut C) -> TriMesh<N, V>;
}

A pattern that is replicated along a path.

It is responsible of the generation of the whole mesh.

Required Methods

fn stroke<C: CurveSampler<N, V>>(&mut self, sampler: &mut C) -> TriMesh<N, V>

Generates the mesh using this pattern and the curve sampled by sampler.