Module ncollide3df64​::procedural​::utils​

Utilities useful for various generations tasks.

Functions

push_circle
Pushes a discretized counterclockwise circle to a buffer.

push_degenerate_open_top_ring_indices
Creates the faces from a circle and a point that is shared by all triangle.

push_degenerate_top_ring_indices
Creates the faces from a circle and a point that is shared by all triangle.

push_filled_circle_indices
Pushes indices so that a circle is filled with triangles. Each triangle will have the base_circle point in common. Pushes nsubdiv - 2 elements to out.

push_open_ring_indices
Creates the faces from two circles with the same discretization.

push_rectangle_indices
Given four corner points, pushes to two counterclockwise triangles to out.

push_ring_indices
Creates the faces from two circles with the same discretization.

push_xy_arc
Pushes a discretized counterclockwise circle to a buffer. The circle is contained on the plane spanned by the x and y axis.

reverse_clockwising
Reverses the clockwising of a set of faces.

split_index_buffer
Duplicates the indices of each triangle on the given index buffer.

split_index_buffer_and_recover_topology
Duplicates the indices of each triangle on the given index buffer, giving the same id to each identical vertex.