Function ncollide3df64​::ray​::triangle_ray_intersection​

pub fn triangle_ray_intersection(a: &Vect, b: &Vect, c: &Vect, ray: &Ray) -> Option<(RayIntersection, Vec3<Scalar>)>

Computes the intersection between a triangle and a ray.

If an intersection is found, the time of impact, the normal and the barycentric coordinates of the intersection point are returned.